@@ -1,7 +1,7 @@
 [package]
 name = "illust-notify"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [profile.release]
 strip = true
@@ -32,8 +32,11 @@ log = "0.4"
 anyhow = "1"
 futures = "0.3"
 serde_json = "1"
-inotify = "0.11"
+inotify = { version = "0.11", optional = true }
 itoa = "1.0"
 pretty_env_logger = { git = "https://github.com/karin0/pretty-env-logger.git" }
 pixiv = { git = "https://github.com/karin0/pvg-rs.git" }
-timeago = { version = "0.4", default-features = false }
+timeago = { version = "0.5", default-features = false }
+
+[features]
+inotify = ["dep:inotify"]
